api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for GTS"
authors = []
license = ""

[tool.poetry.dependencies]
cachetools = "^4.2.2"
python = "3.9.*"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
gunicorn = "^20.1.0"
pymongo = "4.0.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"
interrogate = "^1.5.0"
safety = "^1.10.3"
bandit = "^1.7.2"
pytest = "^7.0.1"


[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \gen
)/
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 0
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
# possible values: 0 (minimal output), 1 (-v), 2 (-vv)
verbose = 0
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: integration tests"
]